split use cases from auth into subfolder
