+++ server/src/authentication_context/application/use_cases/__init__.py
@@ -1,9 +1,11 @@
-from .admin_login_use_case import AdminLoginUseCase
-from .register_admin_with_password_use_case import RegisterAdminWithPasswordUseCase
+from .admin.admin_login_use_case import AdminLoginUseCase
+from .admin.register_admin_with_password_use_case import (
+    RegisterAdminWithPasswordUseCase,
+)
 from .validate_user_token_use_case import ValidateUserTokenUseCase
-from .get_sso_authorize_url_use_case import GetSsoAuthorizeUrlUseCase
-from .sso_login_use_case import SsoLoginUseCase
-from .configure_sso_provider_use_case import ConfigureSsoProviderUseCase
+from .sso.get_sso_authorize_url_use_case import GetSsoAuthorizeUrlUseCase
+from .sso.sso_login_use_case import SsoLoginUseCase
+from .sso.configure_sso_provider_use_case import ConfigureSsoProviderUseCase
 
 __all__ = [
     "AdminLoginUseCase",
